Rename Dispensary folder to Prescription
